requirements_test.txt
ta-cmi==1.0.0

#Test will fail witout
aiohttp_cors

pytest-homeassistant-custom-component==0.12.4

